refactor(gateway): extract media/event helpers from run.py (slice 7 of #54962)
